utils/requirements.txt
pytest
py2neo
recipe_scrapers
requests
python-dotenv
flask
nltk
spacy
scikit-learn
gensim
pyarrow